requirements.txt
Flask>=2.2.3
Flask-Assets>=2.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.3
Flask-Mail>=0.9.1
autopep8>=2.0.2
pyyaml>=6.0
libsass>=0.22.0
python-dotenv>=1.0.0
PyJWT>=2.6.0 
jsmin>=3.0.1
Pillow>=9.5.0
